// libs/colaboradores/src/lib/models/colaboradores.ts
export interface Colaboradores{
    sap?: number;
    nome?: string;
    sobrenome?: string;
    CPF?: number;
    nome_projeto?: string;
    EXT_projeto?: string;
    status?: boolean;
    telefone?: number;
    dt_admissao?: Date;
    matriz?: string;
    tp_contratacao?: string;
    un_negocio?: number;
    tp_atuacao?: string;
    RACF?: number;
    funcional?: number;
    CSR?: number;
    logadouro?: string;
    numero?: number;
    complemento?: string;
    bairro?: string;
    cidade?: string;
    UF?: string;
    perfil?: string;
    nome_lider?: string;
    sobrenome_lider?: string;
    PRE?: number;
    ARE?: number;
    id: number;
}